Update vibe.d & related dependencies to their latest releases
Agora currently does not build because of changes over the years.
This will greatly help reduce the number of errors as Vibe.d
was quite impacted by the DIP1000 changes.

Some submodules were also changed to point to upstream,
but vibe-d still suffer from a bug which seems to only affect us.
Finally, a dependency to openssl-static was added to cope
with the latest version of openssl.