Re-do Tracy integration and switch to new repository
The idea with dependency handling in Agora was to have a flat
submodule structure, so there should be no transitive submodule.
Additionally, as the Dub registry works with Github release,
which itself does not copy the git repository, the preGenerateCommand
used in tracyd would not work there.

So we do the following changes:
- Add a Tracy submodule that we compile ourselves;
- Remove the traced-server configuration as it shouldn't be needed
  (we always compile Tracy in and use on-demand);
- Update tracyd to a new version which does not include preGenerateCommand,
  as well as a few fixes for modern compilers;